server/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-CORS==3.0.10
fpdf==1.7.2
psycopg2-binary==2.9.7
python-dotenv==0.19.0
gunicorn==20.1.0 